# Add fast pub aggregation